app/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cffi==1.17.1
crispy-bootstrap5==2024.2
cssselect2==0.7.0
Django==5.1.1
django-after-response==0.2.2
django-crispy-forms==2.3
django-pwa==2.0.0
Faker==28.4.1
fonttools==4.53.1
html5lib==1.1
pillow==10.4.0
pycparser==2.22
pydyf==0.11.0
pymemcache==4.0.0
pyphen==0.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
tinycss2==1.3.0
weasyprint==62.3
webencodings==0.5.1
zopfli==0.2.3
